+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ### 11.5.0
 
-* `[signal]` Added method `GetByName` for getting signal by its name.
-* `[signal]` Added method `GetByCode` for getting signal by its code.
+* `[signal]` Added method `GetByName` for getting signal by its name
+* `[signal]` Added method `GetByCode` for getting signal by its code
 
 ### 11.4.0